Refactor: /data => /constants & /json
