+++ website/src/components/_pages/about.astro
@@ -5,8 +5,8 @@ import Apresentation from "@components/astro/Apresentation.astro";
 import AboutP from "@components/astro/aboutP.astro"
 import ScrollAnimation from "@components/ts/scrollAnimation";
 
-import Imgs from '@data/imgs';
-import { author } from '@data/author';
+import Imgs from '@data/constants/imgs';
+import { author } from '@data/constants/author';
 
 const h1Html = `A little about <strong>${author.name}</strong>`;
 ---
