+++ website/src/components/_pages/contact.astro
@@ -3,9 +3,9 @@ import { Image } from "astro:assets";
 
 import Apresentation from "@components/astro/Apresentation.astro"
 
-import Imgs from '@data/imgs';
-import { author } from '@data/author';
-import Pages from "@data/pages";
+import Imgs from '@data/constants/imgs';
+import { author } from '@data/constants/author';
+import Pages from "@data/constants/pages";
 
 const mailResponse = `Thanks for your message! I will get back to you as soon as possible.`;
 const formSubmitUrl = `https://formsubmit.co/${import.meta.env.PROD ? author.email : 'adriano.limarossi@gmail.com'}`;
