+++ website/src/components/_pages/index.astro
@@ -11,9 +11,9 @@ import Testimonials from '@components/astro/Testimonials.astro';
 import { YouTube } from '@astro-community/astro-embed-youtube';
 // import "@styles/_pages/index.scss"; // * Import like this if global styles
 
-import Imgs from '@data/imgs';
-import { author } from '@data/author';
-import Pages from "@data/pages"
+import Imgs from '@data/constants/imgs';
+import { author } from '@data/constants/author';
+import Pages from "@data/constants/pages"
 const { background } = Imgs.cdn;
 const { intro } = Imgs.videos;
