+++ website/src/components/_pages/songs.astro
@@ -4,8 +4,8 @@
 import Apresentation from '@components/astro/Apresentation.astro';
 
 // * Data
-import { author } from '@data/author';
-import Imgs from '@data/imgs';
+import { author } from '@data/constants/author';
+import Imgs from '@data/constants/imgs';
 
 import { getSpotifyToken } from '@utils/Spotify';
