+++ website/src/components/astro/Album.astro
@@ -3,7 +3,7 @@
 const { name, release_date, image, imageAlt, id  } = Astro.props
 
 import { Image } from 'astro:assets';
-import Imgs from '@data/imgs';
+import Imgs from '@data/constants/imgs';
 
 const albumURL = `https://open.spotify.com/album/${id}`;
