+++ website/src/components/astro/Artist.astro
@@ -2,7 +2,7 @@
 
 import { Picture } from 'astro:assets';
 
-import { author } from "@data/author"
+import { author } from "@data/constants/author"
 import { getSpotifyToken } from "@utils/Spotify"
 
 // Spotify Access token for API calls
