+++ website/src/components/astro/Footer.astro
@@ -1,8 +1,8 @@
 ---
-import { author } from "@data/author";
-import { socialSize } from "@data/imgs";
+import { author } from "@data/constants/author";
+import { socialSize } from "@data/constants/imgs";
 import Social from '@components/astro/Social.astro';
-import Pages from '@data/pages';
+import Pages from '@data/constants/pages';
 import UrlLink from '@components/nav/urlLink';
 
 const socialPx : string = `${socialSize}px`;
