+++ website/src/components/astro/Social.astro
@@ -1,7 +1,7 @@
 ---
 import { Image } from 'astro:assets';
-import Pages from '@data/pages';
-import Imgs, { socialSize as Size } from '@data/imgs';
+import Pages from '@data/constants/pages';
+import Imgs, { socialSize as Size } from '@data/constants/imgs';
 import UrlLink from '@components/nav/urlLink';
 
 const socialSize = Number(Size); // 96px
