+++ website/src/components/astro/Songs.astro
@@ -5,7 +5,7 @@
 
 import SpotifyEmbed from "@components/astro/spotifyAlbum.astro"
 
-import { author } from "@data/author"
+import { author } from "@data/constants/author"
 const { spotifyId } = author // André Valério artist ID
 
 import { formatArtistName } from "@utils/Album";
