+++ website/src/components/astro/Testimonials.astro
@@ -1,7 +1,7 @@
 ---
 // Inpired by https://codepen.io/littlesnippets/pen/EypPVx
 
-import { Testimonials as TestimonialsData } from "@data/testimonials";
+import { Testimonials as TestimonialsData } from "@data/constants/testimonials";
 import ScrollAnimation from "@components/ts/scrollAnimation";
 
 // * The each div has to have the class "item" because of the Carousel.tsx
