+++ website/src/components/astro/aboutP.astro
@@ -1,5 +1,5 @@
 ---
-import { author } from "@data/author"
+import { author } from "@data/constants/author"
 ---
 
 <article
