+++ website/src/components/astro/aboutP.min.astro
@@ -1,5 +1,5 @@
 ---
-import Pages from "@data/pages";
+import Pages from "@data/constants/pages";
 ---
 
 <p
