+++ website/src/components/ts/Header.tsx
@@ -4,7 +4,7 @@ import { useMediaQuery } from '@hooks/useMediaQuery';
 import { handleScroll } from '@utils/handleScroll';
 import { getCssVariable } from '@utils/getCssVariable';
 import lockScroll from '@utils/lockScroll';
-import Pages from '@data/pages';
+import Pages from '@data/constants/pages';
 import AuthorHeader from '@components/ts/authorHeader';
 import UrlLink from '@components/nav/urlLink';
 import NavBtn from '@components/nav/navBtn';
