+++ website/src/data/json/videos.json
@@ -1,0 +1,47 @@
+[
+  {
+    "id": "VxLTWzQc0KU",
+    "title": "Woodwind arrangements by André Valério . Carlos Cardozo - Les lignes de ma main",
+    "description": "\"Excited to share one of my favourite woodwind arrangements that I wrote for Carlos Carsozo! This one features two flutes and two ...",
+    "thumbnail": "https://i.ytimg.com/vi/VxLTWzQc0KU/hqdefault.jpg",
+    "publishedAt": "2023-04-09T21:59:15Z",
+    "channelTitle": "Andre Valerio",
+    "channelId": "UCwk5prbrBn1NyQFSAbzCrag"
+  },
+  {
+    "id": "l7OpBzoIIAo",
+    "title": "Online lessons. Tabs with real audio",
+    "description": "Hey there! Are you looking to learn guitar online? Look no further! I offer personalized lessons tailored to your specific interests ...",
+    "thumbnail": "https://i.ytimg.com/vi/l7OpBzoIIAo/hqdefault.jpg",
+    "publishedAt": "2023-04-04T21:33:09Z",
+    "channelTitle": "Andre Valerio",
+    "channelId": "UCwk5prbrBn1NyQFSAbzCrag"
+  },
+  {
+    "id": "RHoMcEO6rNA",
+    "title": "Algo Contigo  - Featuring Kalya Ramu &amp; André Valério",
+    "description": "Algo Contigo Written by Chico Novarro Andre Valerio - 7 string Kalya Ramu - vocals.",
+    "thumbnail": "https://i.ytimg.com/vi/RHoMcEO6rNA/hqdefault.jpg",
+    "publishedAt": "2022-02-02T22:31:28Z",
+    "channelTitle": "Andre Valerio",
+    "channelId": "UCwk5prbrBn1NyQFSAbzCrag"
+  },
+  {
+    "id": "gfqnuvNIsbA",
+    "title": "O Ronco da Cuíca - João Bosco by André Valério",
+    "description": "",
+    "thumbnail": "https://i.ytimg.com/vi/gfqnuvNIsbA/hqdefault.jpg",
+    "publishedAt": "2021-04-20T15:16:52Z",
+    "channelTitle": "Andre Valerio",
+    "channelId": "UCwk5prbrBn1NyQFSAbzCrag"
+  },
+  {
+    "id": "R68z66ScnCY",
+    "title": "Magia - Djavan by Andre Valerio &amp; Carlos Cardozo",
+    "description": "Here we are again with the series of Djavan's first album \"A Voz e o Violão\" (1976). Today we play a very interesting \"Baião\" ...",
+    "thumbnail": "https://i.ytimg.com/vi/R68z66ScnCY/hqdefault.jpg",
+    "publishedAt": "2021-04-20T15:13:59Z",
+    "channelTitle": "Andre Valerio",
+    "channelId": "UCwk5prbrBn1NyQFSAbzCrag"
+  }
+]
