+++ website/src/i18n/gen-seo.js
@@ -3,7 +3,7 @@
 import fs from 'fs/promises';
 import path, { dirname } from 'path';
 import { fileURLToPath } from 'url';
-import { author } from '../data/author.js';
+import { author } from '../data/constants/author.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
