+++ website/src/pages/contact.astro
@@ -1,7 +1,7 @@
 ---
-import { author } from '@data/author';
-import Pages from '@data/pages';
-import Imgs from '@data/imgs';
+import { author } from '@data/constants/author';
+import Pages from '@data/constants/pages';
+import Imgs from '@data/constants/imgs';
 
 // * Components
 import Layout from '@layouts/Normal.astro';
