+++ website/src/pages/index.astro
@@ -8,9 +8,9 @@ import { fetchAlbums } from '@utils/Spotify';
 import { formatAlbums } from "@utils/Album";
 
 // * Data & Types
-import Pages from '@data/pages';
-import Imgs from '@data/imgs';
-import { author } from '@data/author';
+import Pages from '@data/constants/pages';
+import Imgs from '@data/constants/imgs';
+import { author } from '@data/constants/author';
 import type { Props as seoProps } from "astro-seo";
 
 const Banner = Imgs.local.banner;
